Run linter
