--- packages/ml-dsa-87/README.md
@@ -1,11 +1,0 @@
-# `@theqrl/ml-dsa-87`
-
-> TODO: description
-
-## Usage
-
-``` js
-import { cryptoSign, cryptoSignKeypair, cryptoSignOpen, cryptoSignVerify, cryptoSignSignature } from '@theqrl/ml-dsa-87';
-
-// TODO: DEMONSTRATE API
-```
