--- packages/ml-dsa-87/test/sign.test.js
@@ -1,136 +1,0 @@
-import { expect } from 'chai';
-import { CryptoPublicKeyBytes, CryptoSecretKeyBytes, CryptoBytes } from '../src/const.js';
-import { cryptoSign, cryptoSignOpen, cryptoSignVerify, cryptoSignSignature, cryptoSignKeypair } from '../src/sign.js';
-
-const TEST_VECTORS = [
-  // Test vectors taken from https://github.com/theQRL/go-qrllib/blob/main/wallet/ml_dsa_87/wallet_test.go#L25
-  {
-    ctx: "5a4f4e44",
-    msg: "4d4c2d4453412d38372074657374206d65737361676520666f72207369676e696e67",
-    // sha256 digest of f29f58aff0b00de2844f7e20bd9eeaacc379150043beeb328335817512b29fbb7184da84a092f842b2a06d72a24a5d28
-    hashedSeed: "f1e4118b9beb0e69d9dee85610b7bde4606a3f8415afad4883e84ad59fe20607",
-    derivedPK: "d2d7a04f06b477907a62a7c5ba0b8d5bc0755354d7dcfed5c4268f324aad7afa3763f3538a0bb6fb4e03106bcf9282b8f06a14d10d2fb325b2dc4aaa729bb9f96b6f93fd26390bcd4fae8ab22f76c9e7f8d2eedd5f4d07061cfde53ee02a9125e0f49336357ee96c0b1494befd49e4771977cdce1a22b445213d8a0af9c6c69c110f6e37ed4589387fbf693f7ccabe46512c2fd5845399790c545da71cc36fbf37f84f810e76d9f6d2e545ecbaea2e182616ec6394229b3c5d2b476f4c31c45c19e76bfce9b101bc9360e400920b75280915e2d2fd3ecabb458c66bd414efb3cbaa82ec03e9f4bfd723d16da088dd0bcd6b62e1cead240eb630b594d1793be1a0dce8e6832e3bdf7767a94f32a096a245a46e8580d137ecc4c12dbd2f3ed7b0fed2ffd9d28e391b959a869401d30e84250ac4d1ca476488430a5a44770fa043c9118701553e8c84bf2c70c0b9ca3e7a4836a4881e754b2a59fc0a35af7f8bd6b8d4abb4d5582e8ec8be9f7ae3a11e32718d07dfc2b518e1075cf36087c6167dc344f91e240f8c02335c8df200c2bad5010f50a1501a32786035c2877ddfd4f74f01fe93d07ed2c4405ac198ca50c2fb2f61e0552a48aa4c177d2ee76b02629764146d98b58af4751d140fc40ca94240f85b43e7bb45a1d9f9728398437c56b9023c2d248011dc6ecb8b9201496212593204722ed2d8054e9f1de6e13c9509b4984a928b003088430001abd0257cde9352fd80cbad728b8645ca7e3d722d5a2ba211a5bebf7adba676a06d4edc7284e522b19412d31a1de6c2e09e0147759cb3554a68bec189783a9ff6fff194a5325db68e1d4903415d4ffb19f18ddfbfa51568b77d55461b4443ea10202d7bc4121dd391915c09654110239b73a4b209f31d86fe50fad3602b8fe9cc6e6fbdb441ce43a6d56ccc11e39e1353c987e5af5f34b0439d3146fca9c8067e4b6ba53abb23ae634f25ded6509fb76293a5fdb1d32671cb87df5981ff2970f816e2011eaaa6ec810a9bcc670d959b62d79df9ae3e8643237d41600df6a3be8c0c8e3ef95e1ce0cdacf0fb2e24e44e3b279ac7e97c433143714c269923acc644634da6abfef9f3a19bbce029f5e49a63e645fbcf08771829e6996faa4f551142a3a2b6bd647b498105b537a9e829351824b873f0acc568e26e554c987ed21ee302e00a3b6177d1779101b0f2945782cc95df387d14425de4180b401d9fe3516697162b119f0d181c03d1449199ec40f03545ce67997715a749085bac6fffa005ba6187130e831af6e41cf2ff2be59cbdf0b3076e6aaa41eb8da4a08920f68b024f41816e35e7e492ecc0b964fbc11a6761e463da7a5943447b44ea8d1d491cd2fa1ab5d0d9f2a3f3719ed3781a961a85dfc94a4539926af967003cc1ae18b4401e448a4da38f91adec2aca6abf445921008d7bdf3e0da818afc64f13078852eeecf8f291abfd5a992d2a82284dc36ca9f3201de67087ddead3f239fb2c0afa240fc451db64d159c0dee372a5f4f2374e22edce0137521400e4831f8164d90991f3eacd106154f24061c6c1a757ff47aeb21421a93626680c61d808c359c14e4c4be92b0fe0a24590c5a406e9e2b6730c458510129df62aa281339b1d117fe38d62f530042b16eda0a3389ea8bb1d617bf0e3e75d03d2cced188184ad89ffb044a91e7bf18b9fc95a48cdb2c9863224df03a9cdd269c2bfae5f312de03f0bde5d6f47c270a679a49af372d96ec84999195ec584cb8a19793f764f314f117d470091b7ff1ac8bc0a1d53b575e6bee6cf76b8eb56a9ac73b8f925657986418a21db2cdb0f5f3b604884ac53ed278f49976859ebf3ab51b4a7d1084a145757f0853579b7bd9b4c0233308edae2848f6c3b3fa8268fcf8e4649c587a89fbc492f2ab9fe2aa8ba62db6e1ad64315d9d974169926f7a45bdb518c4f58f24b4c6b9a45bf78725a1ff218648300383320f53af620c548b91b3f05ee705b3892bd7788c28151483fb20d2bc042f4691a9476524ead501114931b561792f34003be31d75a246ebd1b32214c9b307cc5c01b20f4f53a8b3c1d41b5c62b0754e7efc11e4f0f0d602c09ed00b4aa4c2f5bd07b9adfbd67748f6a1c144f088513b7f963d8be5c9ce5faa0a8bbbdd6406885119d2606de058af4bb721f6d3894487effd64b2b1eaf7edb37572a5b0715e5e1ba0967c1b591d5647a9ce581748ae1c7e86ecb9787918d232a37b4d6f03bd4462d96b839c171581f1c3e7ad689775290bae38e3f18c1c260fb61b23790af68fcebcc94dcdd3191521b1c87b3b7f2f503d8024b9bb8498a470b1dc9ca56eb1708803a408b6d850a3090eee75aba9b407caeb63250ab6e796778b04466f055fa1272df81e21dfff99a780b8737aada488de893fd6f5943ddd5481f6538ca8c1d7853384919cd719d67dd68adbbe5b539c111e7efe95e01cc83fef2a1458d81aeb0c3cd5e4fa3338f10ebe34dded123f612e99cbe4bd2798702c11e65a5e5d615c3e6b6ffec22ae3b8594ee1aa674d44db73a035e7a2214e71e0173f4381794e00fdc14550300ce1fb1fd0a2af6fd57e593d03bfeadf63c4fff0b3f14e9af9d716d4ccb012d248686881f324df220a671957945ba5d8e4411513fb98913608b86a3d17e39125d9fc6e4004d03a03157ca6e4ab885d39edfceba4cd0be32cd94f7a45903d1291ebcabd20a6bd9291228e8b9ca473db08ffe50cd8066323767239d1c780756f12e68d3db383fd9e02407b9c4cfd3d15c1809eab0d3f67cb00d4e994634164c6c0a26337ae6e39c2067b48662d3b35c49e359c0b4bdeba2d167d4e649634e2d2e6413bd3fdf68abba0cae31abaea0596b664d573497ef038843deed7580aff064b900a67646a0eb090dba30ff748445780e88188679e4f1249d8a56f795e31b7ad08cebc9277492a5d774c596ede41dfe67dd7c4a56c1f3651927259d830a4a46aa9b6accbcb58713466f0f14e28867f97df163ea8420a55077b521e9953c2f52049d21530a54b70b992b4b07003829f92ea1a0a91ac3cc186309ba3011aba87272eeebd4e578992a57a9644240752556067fdfb95ead63e29faa2d581851938c76b38fb19f7337e642bedec496eb2fbda65891f9181986bc8317f5f6e2829839fe05c2b44c675a94ccd3296628bc4bb728dd788b6f939da7af1a5e161650931a8e16bbf5035f0769f018c4075d1f107512d9bca93f98c716b93080cc61194dbb9e245132c711fef64442b6d75c3329a9c46b249b1dcecfc8e93bdfa7ce439f66afadbd5fc7b54c86e4783d2c6c397a6b3c482ccffceaffe22ed6c47c27d972a4529cf58c5b18ba50aaee96be9f5c0412ee9bd7e0a5bc0848a519762591cbaa1aaeeaaa06e07424bc7d21f8a12046a7d13cb08ca106984c456a82d9c2d91e188e30d8b6cf62546ff6e46eaed60d1b4af9ae45780524f5bc74e44708f2b995169559153d0d8127841923240543682fe8274b95519615ce375e51bf06dff7f3a6343b6f40c69d6f2b989c8848419c4cff4aa09734c7070ce50d3fc9248208cacb0be1325c226fbb906a32d742eb81be7e92da9b9ac9fe572ea4713524899d70c3ef4d60fa87aa9783e58b043a35e96e436b547a55e91699a21cd76e1afe459bf2c9ac60bf58dd25c11c3",
-    derivedSK: "d2d7a04f06b477907a62a7c5ba0b8d5bc0755354d7dcfed5c4268f324aad7afa3a7cc2c1f05b40ff4296034bfb4360bf3d21b700c690bff605a238fe264fbb356623552a7592f6b25267d40df412c6a1916014918bed88717963fc5e8d83432dee6db2155d4e4bd18f2e9f7bb72be6349571319edfae246890b62c69aee97dcc23316c42b608099169dc20708b404802b3718486709ca80599a86c4b160c430452d3c2240c222420306e4c3866118924ccc61154b0494c288613200e10380dc400418b468199446660a68d92186022978c51406552082a9444449386718414902107899a248a1b378561082d422266d13240099465234028e2a0108b128020134a54006c8a080c13146a1a13041424114b2000213760d98290a3a66c03056663367063463153344e23252122298cc8a4605140691bb98091c8280b8404a3b844a20260c4448d5c3486d3c210e106114c9010889441d34828d9840824822d1c1772c41621d0884d122346411801db24060384800a860510a5484018654236205004711218904b02059880458412899a46485cc885a0b0110984710899258a984183383212b928d8408e90b661c2963060284508a78dc0164a899481db96811b334218b52c584605531885412009030430cb002c2024326440446294409c34485b36304140846442321cc9441ab445614422224672c0868550284823214d9302294b028ae0a4301b112c8aa44011866c21c3610a072410264401474a02312614c2880b292848b4681b196d82824d409229e1204651284698206a19b30c09a13082822910b67118b664d2940840a84cc8a028c1a008e4203114c8458b366e58c089c810321ca70090a8210382304c408d5ca24123962d5280204c426622331101318113a18813a74952384c13b53189224d9198512118690b258c81080eca462493a06d0c1611e4a86064203044b24564044a89a02122c52122b149e1a44ccb9860901472e4a02d4ac6618c128e238330602484caa22510c624cb88914010290c079024481113b26918b848534669e1b26d8c424ac9980c6214710c1240d1842d1c311250944864c40980c60c513445c3409210c40950126e82148a91046221c705ca120a94108862168453b0501b894508a34994203011118d4244641ca630201130041466813470dc468222a700e43422c3464118b6310ab96808073094a0281a1241e0122413976044066c619665d8a66400042e183428d3988154161010444dc2180e9b0209131285e384916090701c228518b380c1320403428044c26108c58013174a8a2052c192444a82509c22084910642308050c402111364ca33624e3c86c04076d20c448939000521070829865dc320d22878951a02c64969001a62c9c906043906c4280600b4302d0906c9b34720c18054a142d13418808b2612422249a987109b3818c1462603222e40486814065c1c650891601d4c80d1a3430032389139989593600e3c42941b8718ba2091393691b376018b141dc98081c12624bb448c9880d1b908c8b100444064a9c0842043492102689a13281919421e2a40da4c6502323281b861141b289d9008c80c431e2004d400042d1362c920861011231080809449464c8b0600b9090c8880908c7910bb71114890104072e9bb869c22686d38884401408a34869cb125292a045c8302d50b811a1146d21c84d02218d44060c0a248c0b262004294912288a5bb0891c134113480d5bb02d9ba429a2346d9a046019026e54464a9a2608923608d04452c18684193312db24490b4502e21026a2a4616092200211480921300c9068d0088623c48542420a80b401239171cab0259416840b47601934494a42202117264a44015cb4614b4421982646510484a314298a966d899840229581a330650943001ab520a3164460a210c2241208416acbc630129700400689c9840820c1691221844b966541846c21368d53b480434092ca42644aa8498b204de22452599021d308440131290bc9108c44928938328a167020a9488c066208a4251ca10c1b095223168d00260dcb40100bb584a0b428e1a00c1482410402921a252cc1088a14a94c402092138165d1a04c5c180c60222e49320800144e59964d21016151824d12130cdb467223034dd3368c948031120869200911c21648db404a03a24c4bb465104385d2342584a0855494053aa0cf0c4a981cfefb3e1a206b7b95f66395c65be670c10ebae3838cf549aea34242b7065b0e75a82a45560a8245c66e6a735d36805f8c924ca9d156ace5eef31b277bd940fb4701defb0af731a2ef455433c6a85fad10909a55ea61a58483c0a1ca46117f1720420dfef998fc894cd8a229901d426eb7274ad4d5b04fb382f900fd9aaaba7302b883bacc9ed092526f133a632c35967cb3edcc8fe27cc75abd05dacb212c175fac2d93faa0c11d5390b89f89c50ed50991b438935339b3c204d7f8f6ed7e3215f1124d3c821bf77f1e924a87cb116a461584a24e9bcc3a87a7ff70989841874ad7851596eb95b1716d93ebcf2bbbfe1d41f36b4c1f6499183364c340d35acbf5c731b2b1e1ae43c9936f376821be2721d3413d258ba076030a68c44734f56e262ce4171ae8480e0361f0314763f1614aa512810102c1b759d193b0986ad282b6b26a84e69bf822c90c2a82337fea20746cd0359ad28874599d191c1864649eb0d45f1b7526138794bf0519930d8e48edcbc2510fa14115ea4f8d55e91dd9bff53271067120629635ef7e23a291621ab030190c4a405cdf4825e09b451144ed41e7e072a608015c3e6d2256331a65bf4a387d2340cef55887ca4247f5523d50a4d3afd49072ea89a322ce0f7b27c11f091716cfbd2f58e16d2426e19216f8ea73b5aeb074ba0da97c33a232a23151742e6706279b8d4cb10cef0631c30fde83f0a816dbd28de5c4e30dca11546ed43718805ed3bca8d9f2629297988e6d2d70fbf4e7e4e3aeb9aa220b238abcd738d3871f2a48ed6a62405c1436ab97575be4d5ba6afee621c2c2cf4f75daa94daa134b816539f48d1357b2b40fa8bacccee8ea7516017a0fb931f2610d5d91d669d90814e80db314d8b84b250efc7ed82ccfeba082435c361fcc5eb9b7687cd0f0bb20af8db9faea63fbddb84b5d7e7bc9af5424b2c209d87cfd4124d170d72b9d3f0252c6fb95d64a34da6b0251a12f1849b254b6bd0b7f5c7eb7cf41d9d98ac60fe8b5600655e91b4cc90737e6a47584194a1f06eac9599a8caa261fd38b55bb3cd77e4d6e495bdf776e9d9021a52e756727de756c5bf584994fad7b1628444c275d0606363a9c537c55759b1c1e74297aef9bf484075818dd627942ad1e654da45dc4a6a8c969aae00e57da69b012744f8eec4e08c05ee68bfb5917b2b1231190188c69519ccadbfd456a2c9488f30743c082542ba0f59e633c401515bd17b369099b3c2cfe605d4f7ae8e0568535f97f587d862a20a118b9fed6fe30b32328a2f6a52fbfc4bcc2b9350d3da32be72052cdfb3059d902cba702ca82c623f0b89978968c419d32ac1f57d45b83cf0496e2166d62a7e7711d68a755ea2b601829cc5023cd6e9edea2a3bac1260e99adb696c9a8d7ff37ef0724cc04d21a53c6a8c6c8fcdc4864ec46993c0f92987a6a95e85a8f743ec8502c34488dac11a0e0bd42c0e50d56e392c49603298ecec6b7e0a42c6816f1511345b9457d4300c1adf6df287866fd11f4ee1a2e16c0f6c43382980aefc8eac88b76cca070f23a29e290351f01d4f17362209dd3a3e89be8e75c0879863760cb912fd4ed47aa89861e907740b5cbd2bd813048f5918dbc104bd170c440f342c8bbb51b45bf904793a26f5827f4b14c1837be87da8d25b51173fd6b06b4af8452e8dff073e58c7f14221e7900cbc11635cd18b046ffe2017956d36161a6d3ce458ef4b8210d6b7149eae7f4a57a3aaed988624200ec29411a95b50c84609f8876ca4f4ba8f919d9b5f0ba7f96b8d04036d23c20a9231324def4bba24958c0cfea232ef609e14f89b31760354f79fcb571bc32b14965456b9aa1563169513fb2c2f9459016b11e8b806ae81c5c2177f5a4b55ba518ff80a97e4e29465a3596eed68fffd016239fa1ce7d860343818ba123dd082d6099bb7b7c35bc34ae0fa9aadd53770ec270951e45d5cb52f4c17bd5c097debaafb376c531add50a985b21cb1387a0d2b053459751dd6ee74bfae9ff8d1dd9eb998643a7093224585daaf9f80aedda6648dc3f11a1bc4de4a29a92972cd2f117f591a1e853a161b1d09a083af6f4a6272d9ea018fa4cc18b53596f713c1a8ccae68f5ee166541f5743a5e24732edaa59ade18c1a645046773e570552955c892e26ad03c13223f02eff90753dbe28b19f96cbcee7bd4f4f977731f6e13b230bf639154dbeb32044be1cd1cb2c57c78eb44b82c5eedcd4adb57410a85d078b83451c4900f8c7d0a8487f322068854f1c4bac67b6ee85ffed841c34372f90075e9675d0255ffcb31e6cadb1431af318143c0739e67d5fbdbe48f733e3ff172c3a66de9906fc62245407a209b8429b8c53780c700110e8d89a31426d243fae5ba847119d1d3819cc9f68743374302346669a3b4bf489e77b1e061096bcde9b63b1f7acd23a37221bbfc064a6af67c44395b3c52ed57123ff09f59720dbe77af462b16ef57cb9433889cc5975c0931774300476a2672301d470a489673217635a5382a8272f741baf15ff2f09198a194d9b586bcc524b6c82f822ef67181041084176386773ceb8ef6417277dd8abd5cda95e00878c2bbd5bc801bedaafc41740b06e7e2876873f7e4dd84de7c563e05d340d42869614cb612c1a495546ea5006a2dce17f3d703c62fb24c62a5521ad31dd552c45b8da21a97293283322ec4ad7f6975eb2fa271717926d344fda7dea5fe71401f2023bddaf802fef794bcbfc05c53546fed16168b2609027b7cd55faa487a6fba4b22eb72fdce5195b783924b87e5c7b958186305146f79aec1c2e82ec57bb6c561417a91cd80b05bb2be67717b885a8f881de9394c5610f858c6e0533c182dd4a9bc371c580e4a87c53eda8f2df173f4037c662296be6876c3b290b92dda5291d6468e9f193fe5f05f5e913f02ce4ad86457655cf61f531f17c66e0b63992260d0167fb7a399f03bef03f5d4af9c69500c438e45a4e7bccbce6cd08ffa5f67c7fe3bc93e2c5338147fd0fb445a5c77ff87574f645c07b9eced57f0a71acd8066daa248e27cd768476a86091689584fa10ae747972d4a005cddf22f404001f4f008fc4b511e00f6a14d942d516cd3e1924ad0c9ae321faca2e94b3dc3dc9e6af22a18dc9a545db3681c6668ba20ce8114a75658c438f0702a33dcb5509b7dd987cf0104a1927e1ccacba29159566ee059f3704d81a28771b7f5c51af330b50a04b0f901e10c6640a27ea9eeb53c895fa381ebaff0b5e9032d7a53889600303148d1f69d62d1d54a8903b6e93d339d71bc03a80a1d206e4217bfc4b50cd6ed15d60f2a2ec39bbc8b9bf55814d1f78491fa42a42e54fab4dda91267c77b05ca8593d7d605a93c057abc292ef6141651dbc33c064c21799b7aa8c383389b95c3f4ff6d7ea6e7db11236cc3a3b88bed1e3514abc2ac7ad1a862a3cfc7340608cd041b944bcc48ad1440ec26ade7910954fdcdfb23bae82ffb675862c00bb2abf689f644a547097752ad82f9125f3d52520cfa3ad7e4531320859e6865238b67c87d3e913ea4f52bbbc48af499899476ae2b0c317c89357794ecbf22271c05de454ecbc8188cdebf6a5119cf5e4726599ab5c0fe835cc9c140947840b4a9ef1e3b17de35c36cb0b5024fc5731d225b091adef43f942efccbfb437056faa66b4ceb0718c11b201cf3f27bdec7be8c92988b39ebb4266ae25c3cfffb14c34843a2971fb6c9250a7fc5c8692cb2d3faca049353f514b1e0f1ed1cfe8f16f3fef01e744d6aaa77c35ac56f3df460c745ecf1befb3244d85608910dcebb0bf504098e1b0c9dcf7bfff521b738c9585c0cb7e0cf4597abec48c54b73ac7cb5205e28a9f6691cc25281036b80078b8841242c7ab5180ab47eb202dae22e678a63a798c759286961f9721a8654fb2b00c7bc474659b9298d380f40907378e6b9788d02845a1bf7c45d73813d05e143f020ced61eed0df51476d62a5b44ba6d26cd8de581babac9c1b24f6da35791fd5ca644a6cd6c0269217b8ce7a897cf22968eedc8e9f8d36ada03b8e66c77d000566ba1b009786cab06beebde3026d9726c9ef58135809a8163aa9f65675b53571d20636a29988a781b9489ec7bc582251cd2a63ad96e059e12e03d3f47989fac26f5a9b562849f433249aed4065daf22329441e5462c64f38a2ea7b129560de3007a7d22571d5ab3d2d7e7ac05909c7592b40aec33d73133274922ddd8dbf152d8a1d909088b3203ce53e7bef3f76f7cf087c95a82421262eaa5acb4b647f2dd393026dbec206d234c9ad217eac4fc1c72d969dcd3a92fcf25f26bad240b14b461159458e3a150a1a772c86ddafd3564c078a34e09d09ac1662f80bb123a73d09cf341ee076ab0a30c74ffadb94d9e6bb5c9681b50d188041fc162d79bac2c103efe3931b4d0d2e6c23b5fa47ec828da9b5cd9a8f090ef192b7b574715ddc7fd892bf64bcb31f417e3ac78c793392912974650ac56602fc443b59b778d370cb3d5d5a0b1ea4fbe80382cde8741352201e2b5381f0c5fa1c3d4e975cdc8b548e749b4df2faafc8be56d3e4b8c575cd6e130b48cf3557a28389094164363a03368b55390d046cc58a379948b0d412137264e5ecff4eefe0e544a76eb3d1a1a0b9af67e4eb8b3c801ac5b7fa7f285dd910e7417503d9e47d54e99896664ec3a3535ef6373758c802d929",
-    derivedSig: "b420418a391f820a2417d2d94fdb4945975e2ded6c179af1ef4db5c9e7267fdde366a457d0e819d05eef6c984834104dd85f01c7aa136ec15ba5f39ce3dceea920879050c3b553e7e7331ef58c9c254f05491c3ef95bd3cc2f93770a72380d29d5d9cbfee08d55485a0105c6150b13fb085188067566f893242f492cf927f84edb6604712b216f058aa1bad64977ea3e2af3434d85d81a8260b0e4e73109920b9f1dbaedc5e38e5c0bcf8e7a322ad7c0d677ceecccdac69b64eec75d5d6b4d23eaf91e7ae2ef166a5770fe4c9b01c11d011d10228711d54ec024d704a2848df6a68b1eaef4215ab150ab46667d29051374372913bddde231db6d1295c3cbb765843ab3b5f0a2a6f7ec49ee6095b6d4c8bdfaeb44effe386b15b8a06e75914a85b6dd142dfa8d2839e6af5e412bdfe5c89a75b838caa7fa91ebcb83a8a7b00e3acd188606c06884263f8d9142f0cfb55d94aedd820676470288e5e262d0bbda0b088177e8ac7c53ea9d0f084a1a863c60a7194b59f85667ec269d798c3b14a2c292e646a11a911214bb8504814aae1d0ceb7aa08d0769fa4de1a05928c00d4d4a24182a79ef2dbaa3969e30c03f9d1cdbac3d5b2f9617ece900b2054e36894a6e1ad2da87e699e82ce2ce43e94a396d2f3088088cd3c7f7689f29b5ce8ae4aa6ea769f32bb3969ea76e117d1983991399fb119ef4900a1124f70d34f2a008998e209d1699fd0d3537d0f51d02e390ab86b873aee68eb534d2c97cca7a301bfe80c01c0e422e6ab1bb5858ffbf05bc039000986ac09562d78ca53f8b3195c9a0a5ad32e74ce463982078baf0ced9b7f84b588538d2d65465dcf17015f0977247955801ad6ddaff407107a0575477540eceb2fc3909104ae56e6b217907e0be9eb04db11e37762e5c9b2355e0ba7e4034a4e8cc442db2cb1a1c075713e3b499d52da1fdf1c80b5f63cdd28fb6f2db1bc7f85a6d176424e40c33fa267112482022cd6877252409ca98857163b5b2ca6a911e21c8e9de554ef466196610b7b050a790f69994e8d250000d9ff70f2fc683ea15ee1377a9e33e789974a5bc0457a105c252157575bebf7f67e0e2f920d23908b177550bb694b5f86bb8eed54960d09acc0c89f2371c7bf2c0c06aae820478e7be1b193ea459487930b931541d10bdb0413869eedc18296ee7420abbc0761dfa2418a00cc651aa9a770b66837824493afea5359887173458cb7c30321badb4f411d1e23277faaed73ebbfad93b0f9f1d4af4506d402beaa4f02498353199f28d7eece77fd5a7006e470a0c195cc4f8bd09a618adc9dc599f8d03e05af64fd5f1870f703a1425ca0fc8e8eecb4b53c482300b367b17d9875eb4438e51a7b65f98b62c2d88efe12ca5861429f4a64772b62d3aefd6eb80dee13f9ff407faf3b57e29569d4964f3b7eb3cb523128a11fa3f4fe8af563c7233942ca06593d9b516afaab842d07893321026bd94807bab683b5e03baca7104006fce348a9d2289e17095239874e301ea3a83e3a25322878e280f75fcfad2fd489d6ace1647d5f6a39378020352e2e3d33cf92aea6fb571eee7cd5ffd48bb6302e2594d37168e7aa9457da759bc752396fbed71652ab7568c5e5f539ba2e29b9e525f39618d0d3be143bb143ed99bf3faffae471337c2884ff6ba72ed21030f5d011eba1aeae8369a6bd42b7458bba6896491fdddba9506a890fd556309f1c79a22f00f4edca7e9913f39e7cd4f9cdb6c111a4ff04afe7578084c83439374ca93a2637a786fb5de373808b8306390661e8c609c838eb4ff00ff0463721670648eb7f2e4b8b4a5936506b3a001fa4e2cdb6437684bd5c97b5a365ad226e7d5e61a8285db3b988d76cf43623e4c2cd98d27abf398b2e50073e12407e3b0069de34455281dec34ee9bfba79f6cefa607f3863afef90796b97d6e94db73074b4505d049a7731b84c8dcf5f27bcef9e48988e8e56bb0a0ec55cce4d2c4a934ac29f5181510025975e18d9b7791d77ce599dca80c560aecdd36cb6007267e0c7a8bb9513fa7a6be1cd42e51f81b3f251e80c4db83290fdce7b946860e0f1f92371924710f3fb94de620b1470af3ce87e5badb09a7fa3cd6c637583d90740af998245b171c41eff462e98f307c3a2606e0b68c4dbeafc8617fff4747bcb572ae83b7bec89aad32c56b29f6ca404b653a4ff058674ccd1d812b09e6c50d261c0ae72978adb59fb3d45a37a96a01b96a6d79c9b4e6d036902319305e74b14069d03696b76a41eb06e735e2ac9c5a2bb764c1d3e8be0721f97e3ca46688ab7405b1fadcaaad3bc86fcd174f7810a72daf9ca4d5303900af90c1b8348f62376de09a0ae335dfbabe399103a00b865a52e93592af97030d3d9f09c9f91d3a72e11c2c120a097781f5fb0d33c8667c5a92eceddda4e45e93e0199c4946863d55121986c67d541f34dd4238e0a3daed1ef3550a00f786ffc3435cf57427d4b968ec39e420d2c05aecb2d1c65a0e35136ae8328b3959bbdf9d19a393f4659f7dfd10c383a2147bdaba478a617a5a67a50791e88ea679b4ada141cd2a18a0f9589b47b3384a1332a1e2c80988aa482525d74baa0560a8f64be3f5469dcfee8c3e2edf6b3499300b6d8da7fa66db734e8fba0634d78e29e23425ac1c85f306236117965b9b5722abe0b1796e74d600170d5f0bc748cf54c8ecb1d4931532aea9cd0401e6de4d17dd1551943bcaff5a5c5255eb9d2e98514583716e9ea9fb164fc7d3022012827ef84becc6afdbe1b350ed6ec6f9100d76098efb34f16d9dbaea2f629c224b4fc9b62ddd49653a119e7854d68698bb30d479603d60c70d6ef636f6b37946c95b8b73dd7725884d3f4fcf6272065ecdfa245b3b4b35af068e4edd443330536bab976460fd7349eff9fd44efcfa57e368fc9896dde3a51ccbf9550ca2566e0014729872a29e53bf8ae68371a06b7ff26744dd1fa3974e176f42943e6d26f7c70763d9ee9851af79155ba53d782569e87b5392f08929c109c725acc7476ea79d9f9dfd0657181d5f92ae53556648dd00a9683dba9a48cde6f5d148864333ba8fc6f041eade20a68332fe5efd972845ff0c02c672924908e2e62c468504d5ada216afd3d66ddd601a972c962c75d19a4444af9581b26362369048a5e81baf83c540013562e1892b40da4a296ae7f2a42551d6cba4a8fa88d3ac5574e337533fbde84a8969d98d128daacb681e5da98186cadfd4adefa04d8a2dbbd57bd4c6646f49612eb1f2a7889b21999f94ec56211f14dca19cd12fc3ebcca272bda1eee52a57cfb1bc9f04e4ac55703a233f75c0805105ee0b6a8475048580345ab7cc98a544ee41e8bc214d0d5874e4846943b258609f178a81590f01f14b9673a9ca745064d261b2adf173df404623ba58f5820d1c3e85d52d499a67e0c55e45993519d7bf0bea9b13ab52b9bc2b6ce02603e7e20ae7f073b01b33a2360e5d152803419ccbadadfe221c3b0a84754ad7a4f9300559f2c0d5f55020424d0603f1defd89e0f098a800594b356747dc4beddb7b3b455dd07c13dba063384295a2cb59071928dcb04c9f291048eb59a6222a921af8debdbdd83b7e0c91614b52fabb62b82d4d84775b24122636809e6207d9b00103bacc1a843e899ca9521ba2b32e247f6b9a27aa4504ab2ba1b11174ecb1be8fe119419565d7add3af11394fcd3d0e399dc5be7fd22cdb5f254519c939b7dae5caf41746840f41a57ec18081afbee09d52a3d22792e5653c13d03015c5245a363761a6e6bc119822dad3a8e94ad91b544beb99c1114fd56b9840d27c0e9d9903b2bfbc05a4909cea3e694795fcb7ce310f5d196ed6983dae4584fea0138ab94909222f0f5e30042dcda79a46a1971b1836897c488d4e381143b71b2d7a6a4215b38a243655f65ab3e39337f6e102594b668e147c88f6dc9c5851d3f27d63dc949d51c03a164e21edf4600182146c27c77e9203008daab3508d9f05535f5dba476542ca37701571040c169ec22cf81da72107466e122d2fa9cc0e31f3ac129749cda83cb5d81863efdda2df97902483d8a3c99776a84229c44ef40a3c75503671c142fda9e15e76177c6512f1130d952b472aea7eec3863f45a0984f7e61c316587bf67f6c4e4d0d6d619094d8a9f9cbb7d7a2cbb9ee17ade2ab13bf8891b635b5a14c1fe8ad7a4ff95077f42fee11e563bebe78676464fdc1808317609bc02c99d9fd970b53518888a5c3dc1e7a03ce233c32f711e9c69a3a94c88c5184696f0b0ca1d50a2c885cfd678a8db4b4f3544ade59c38ec6516eeb5842baa934a9be591f3c4b685cfc2164e919a15758b56767f72f26cabad52d6e7dc49fd78ff678b040f15a4de02fba9192920a303b238149828c328e826aa61f53423afc2fbdcae247a637da69b55a0487fe6171e508d2f514f9b61ebfdea9800e56937cac9a50f0199fb7527b7e82ad7faca5facd7feca504d8e53ec52ec3819469ee1f60a27aa09ad347364976046e7808348556981d0948d869b01497bba9826370978d394ac04eb34af3caacd92e7bc609b425fe8632e0cb5c948d3deeb80bb52a1af2d6b4866ebb77a14bc16e421cb6386ec1cd08d758eea85ab1bade56c7ffbc0e02f1ea1b4af59a8b97832e2ef20e137b705a997b01cf70c0488593b78a2b03111d398079876e4d78eb09f0ab6d316234baeee5d16cc5ff27058a355b1d1d446fec2e88a2cd93e2484d9a0acdb448d821f5e5fdac76483f245875dcf39ab7b94313c5cf5a0365b7bdc5767d54fd396a49ec4d75414a49c14bfcec5fc675cdb0cbda22d967b4c04684507641139f57908d20634524b1d2b9e69ba1da033be708b489d2aede6f62d7f0c9cd6002314c694726a5aecf9e9cf98d2a8340c9b183f1ae341147fed7e39fdcdcc5d54c7eb973ca3984b54dbcc61e332df8abfc20ec323393e4731a5598e85ab16faa269a0ef035c3bb424914cd75c91a9c60e8e976f424eae9e84a1fbc2c88bdf8ca7bd252dc9767612ea6711d497bc3dbbccdaaa7e3f934e676ef3464af76197508843951af14c212d2a8ba47d7af64fc6fe415958db233475471440299ecba57ab9c434dbf4eee99995607c08c00eff4de386a532213ad477eaee9367881f7d93904b6000ef82c863f83057f5cd1af1d7004cdbe2a4b27d3ffdaa9fd6ef9b3c9c5173fb9231d50ac7030ea0be204b7cba41a1511123733c835e75ce3748bec5b70f616c32df8d29007ad0d878880ef75c46658ad9cb44ae0075ee0269004ed58aec72a189a1f51d597e2433ed7f29449d332894eab57950e382d80612de14d5aa2fce22a2c8324167d1990636c15fa2dfe301172a973cf7e7cf94418822abef7223223d4496ba91ba7d3b50cc072ad1c3fe729f1aa61a41afc224c9609fa745c5f9d086db5b6bc7a3bfdf6e67e2b0ebb9993fc1ee7b0a08757845d05a9d95e9689b7ed53b5870cb6a2596d5e35cf9f590401c1ede237969b77adb833dd7f1881f652495b1383faf7b80aec52dd6175bb65f0dae8488d143d6e99ca05c995687c703723d26065582daf7efbea7a43db054b61ca4da99c93f02f115d76e40cfafd7ed7cfbb37cef31ee7ac700aafe64f08c6a90c6de9fe8addfbbcf21640c1e35dfde735dae70885c06b1bc0cbe44caf6dc0b5f9b66bbee55c1c924ed2123a73fda69f03318942e045064683c2a4f3ac16cd609abfc518a0429d4c00e38cd31b4aaab3bdaf2c0a7cf0aeefc5010f4daa3850fa6b7947289b389932235eb9a9ca0ff39d1b00df75f342aacdee3b45fc872ae7b1ecaf782edf7c49e25f6f2b230be9ceff0e4f38b0648c2d0afe058741f1f73817abf0f8cb3d571a6d256f5aa75d4926f86ea529cc16a1f7c748dbb21eab17123f37e1d3edbc306716694f06e82aee0163e4d089ca44154b75e0a466da488e9d9618837cfa30e8745bda1484d600fb82f94ac83f1439f956bd2c217a1b55a6bc84f3a89728e27f241946241b235f39872f431be2d86b358cf5aa39c27c3387a74327c1458a4197c79fdb778dbb8e003ba1fe0cb8d7f5a4b59581bcef69a6a219324575a5ce150ad7d6c7798eae7681b1453f947cca0743605d79bbe0f0164182473c889d53c62962be351f5a09cea50cb8f2e96b33c854027c180a91c817fd9adab42eb354ece697289e9ac9dcc8317e39a5fb73104e8484aba3cf77bf88053a257e8b5fa34b6eff2cd8d208ef20d5c7d56319ffb817cfd5ed1bd104509e81a0a590e917bd9eca2a360a3f612ada44596624726501b8e8925fe228db6ec7b287b75b3486e704d7cfd44e4c4b78301fcaff73519827c669f1f4bf21c04d9674ba4bebbfe9b13acdd01acd975e31c6ab19e202f7d733fa6d18df68cde08863ea051e47c0a58eb695ad394e26a39b0aa69439ef9c7d10b0544155123441657e9dbecbf4133547555b6a72364a516e9fa4aab94168979cca294550559fb3b5e6086c739fb3e7effe0f151d3e547c83ba31354147a1d7de0000000000000000000000000000000910181d252d353c",
-  },
-  {
-    ctx: "5a4f4e44",
-    msg: "4d4c2d4453412d38372074657374206d65737361676520666f72207369676e696e67",
-    // sha256 digest of a7b1a3005d9e110009c48d45deb43f0a0e31846ed2c5aaefb6d4238040ad4c08794ffe65585c13eb6948c2faf6db90c2
-    hashedSeed: "0133968a0085e3ecf0d21e0213e64c37974a0dc1fc74a073049f23a4efec04bf",
-    derivedPK: "6c5c0256b7e67cb7e89b75395082c19cf08a31bdb0cddaa0dbf9099951f1eaba6b78effac950c20e2f10a07e53303e56ca0228685effffd57eb45677935753560cc2301fcf56c73a80684609704b96a50bb25921492b1fd20ee2dc796d4a9f0c6d1b21875c28081e722470236636ab762f8f7b3451f1c5d1865cd6bc75f9e063a93921f4c263413956931d4e4b951bf0c0915f130f32e9334efdb6df5e761a981e0be8708c07d4948643f85862adecad1008ddd0fd166e7933230d8683944959d99f22b61c59542e90caa92c43a1dabda2259029fa798d2aa4226af07ebf1de625603d173833064141bafeb367f2a5df2c177f229dacc119c5a94649a897a01289d4e0d4ace80868153c96fcdb03b4f3c8aa8fdd6b8ab74366322454e400fd2301e30912f6dccaae2647bba0f4b2e893c50f24ac81fb2b9f2fd6ccb8f4e6e1438054dc0284650ba99320654140da53bc57b927bdea089557661056290d826c547483839f8ab77467d959ea0a4bdcbd955b96e27e1d775ce35550f5dfcd8c4347609b75f8fc9746ceecd2da06f36803988a9bf902ccb27c1602333cdb9ff2e4409a7993fc5b54d5d57d1de56564e0885f1d71ee7144df744b0960f753d973d832cc1b7b761282a366abc3a42866b26b8c84518fa05d6a152f017e9bcb466575a470db29f5d7451ff3ae88ff706e9b71e2da68b4eafb82cda58b7ca208685e4c4caeab1ba4a79622e85c08826effe2224f2f16bde95b6a7cccde9b469e9b991e4ea8fc3f1eb54a8fa719af71d321d7b893e3751173f9ef2c48200695f1e29b17597dd5450dd43b3a04042c38f2c602c88ffa4e3e96f1611a43ff6c6734b63f685a9818c9354b900c2d2474afae0728e69e0ba93a9d8c782011fb51e7c6728fb940fef69701a9b32e8527c0a47f8a70f896f5700219215e78cd776ef5079fb6ed9caf20f93c50e20a6e3e7bff8c9b7bbd800e1e59139a3ab3aa5d4018dcbc7095b0d017b77185cc5400693992b1ce633b8d0406b3ba97573108cff5a897b06992bd6385ab5f3537f20d607788d4f0f956103506a0337295915fbe02851b9fcf32e8013a9b433c3250f072571b148436c898e6e78bed671c37fc467491254f68debe627b114c5cffeb1935312fe5abeb4ed22771a35f4aae9b64399908d4be447566e4bddb5ebf2b3837a9b99a7dc59cadba535a77814a43000464cb35a2998688fb2ad2eb66ed367807c9d92e9d776ad2895fdaf12d37cd8c61ff46d027d9adfc0316ec5f3b52f80af2614b18f1d7bdbbda4dc35ba5363d960399042efc1a1fd41f860ed1422ee84e73dc12f1f2eb895d59660c11b1edb72509e5b2212716b055fe52e5a58c3e9b5d22c6d9dc4ca2a1f205d65bc4f9f24f44c72cf42d9b8b91c65858332bb24324b47877aafa523f880cd4724ff990cd2b079105071df7ee90692f43ce23e71a8bc0745984e652b65e6013385e81d44563491f546b0064e9cfb7643da4393cf6b3a7afd4fec232016700c65a8531a88e213a9ab5546255528fc0ac9992c46c1f2ac3192adc32423dbe052dbe33e46aa6d369717c74da5a8eea8e8e281b1f37c8352d2879e4ca9b8e3e0ea3454744ba8bec5434cd3ade1c783085a3dfeaeb90703cca8d0b9fcd9ea063b2ae88ded6b84f9a1498252483a5c7583756631da66bf81ec480478204731baff7f2f2e8cd6ee8a2a76d1dbaed7a8254eb002cc18fd65b57f7982176242e546fda668458378e52eb25ccf149ca097381a878643a34c811d4391474d5d09f5d4317815fc02095cda18372d72ed60d35b70c6554bb91dd8ea4a948ffbeb1bbb130b1109cf4d1702070bb8c9572674fc1df859d9aac67a541af321aa518b1d0de130621141f6be4ffc9226afff782cd2ebb9b1f87162cb32c7df2ee2931a55dff603dda6f81bc1858ab48044c1e6c45c3740c8386d8c357ec0d9dc0ff9ddd7f38ebab0e9f6e6d0f23c944a059109d332e35d2b786fde3df1975b761cef5d52df5f837315c1218d9003ea916f3dd98f926eec8781dd92e726752095e2388cac1a3c8b41da90af4942f8e734105d3b8cba0f310dd38267d2d53e0aa00780efcebbc267ff9bebc980918bdd1b4f42655d0671de9b762ea98ec55fd5d07bb102aeec85903f6653f47ee4951213170bbdcf35a276e0fc4a10df4077b14e9c326890fc502dd5422fc601ad76d45e2ac2f9cdcf1802b4fd2f54af989b0c4087feb032dda3fac5b292af1fbe53013acf162d0a7cfd908d7be37d78904161a04e6f67a63350936ba9d0a94b9400777a3cf2e5c78fc9ea4d0b2de1bd80bed2f0b2201d78a45f37099f912da6698f6bc1eb5a98bf11bd7f3a06fe95a7211a52410913d5ba04d1f80ae24fc499d380980b324ea617def1fd7b90e15df83cd0233f9cbf603aeb41f8f2f3d796596edbca55d385e2620542ecf4fddcee771a69642e7c7bbd2b20acf2774f3e9b0e9adae01ddb5925350b9977ea53071b0db3306dd6dd7c667225d540595013398205381501203bb05bf7c44738936014e1ad5214e2f26059d11f986bdc234f9da8860548e1ede1ca0bf87a010b503f2346610137bc7df2295bb5de4d38c5dc73a7738294232cae9effb82e5a7f012f7b15e223925fb9a1b998545efc28b76a2d9361d545d603cbcc019f70ff073d16c5e2727f123d95eb42339ebac494ce7f28555aea065e6161486fde260a02024f15af6592892934e9fab09025c0ac606e0491e965b58d1dcf60058059e4726728e037a698cf74bf2f2f0f4f9bec51cbc035f32eb3e617cb892af5406e6f3ee74e177a1bb07d1d3b5f80cba842d6d2e7eeb5b2ad7bedb0d9f5e6255033b1bc187064b19b72a2eb926961be1441426200ce27893e30cc217e0a8a9c33fdc22e7c94b49ef094443fca1e528035ec04c83f187686d22862ffcd9edf223d2d1d86da65a597f9f1b784395541bb5d960c1bbc34d0c15404182b3746f875b06bd5f4d41ccb58ed03de6df566cb85595b21ce3bd7062a515bc365c925965e8800263731618f70258a731e41eb45a7d8e19cf18ee8ff0eabb0dd7c89e0f6aacce42dfeee14704d53105f81b627fdb6399c23d330d8ed169e4a76c3d4158016afafdade3b725bfbfcdd8457eef2a8f8566908da236e3a0b35486e0cc979bdcef27329d354624f346a308470d813b28d204b7fab4613243b37ccf92b347390d1da0bf902cd4803a9aa54fd46fc85ff63cae54ce45785aff19cf6b1cf7ada4f3fdaeafc8989e7c9221efc0d5048c6b873bb9951729f92aa097393726a1e904460eb6344a7f4c8d680c0bafdd2f80ac559d82b4970153b87fb7724ac9c92a8dd35c8ebaff1f3d99202d8456d3ed4fa8728f1d4901c883e7cb26dbf804871e4170b6e274c1514d5a64b3c8f4bc70ee41bd0bbc44c1ae016e9bcdf6e71158d37623cf40a64cdf9fb4c230ca8680cd74d49de200acb00ea38117dd8cb246d421b29c355de0cee7b0de003e3c55821a067d8b0b5f8b01a4e5634e48131578a89870e059fb4794d45ba0713179a545a2ffa43d92f178dcf7c6f05c80aca3a6b16419f8d9f6d118caaa5a59535b3211b8ab651de3444ff1205c38df1f7d6b6bf96e04a7d2141a248e0b39619f3da001c437ea752d9c1605d4c11950ffeef401dd3608f368b4e365e9",
-    derivedSK: "6c5c0256b7e67cb7e89b75395082c19cf08a31bdb0cddaa0dbf9099951f1eaba609726e47be0f4fed3e5f6215375c755e496237b0d3b7ab21b97e4c34f024084d8787332cc930d305a4903b76eeb6b15743d2d80540594c210772c5c48437cd695980e4f5d0cb6bbe40dee6e65f165e0ed3076500482bf9b2daf78762b6aaf8b12b8702421899a2890a0185200236e10c445dc28280b812d62185223240412158809230944a8118912265a48442223441013462417505116505aa20110303223b11043243219402001860d4c4848d90684da142d63204080220a191489532666138664033265caa688144312a43605db122a1239400b1705019364d2c4010b37451c1609e38690ccc8641c354402190e0429854a4466081702c208701042701c456620844809c11193126652468904b74812070559b8118a4492c0160199260593141118112820238888b08d53846d1347258a8071dac28514868dc0900d1b27884b408600a6088912495c326091c0911201400b26820b022008a700da360258b03098b848108749c3208598b68559464c84280acc88280a08064aa82509339063160ad9a6685c362cdc402d141720090606cc4652cc4006424605cba288230002d10489d436302284688c262e0897611ab26103c89180864419c8649b020d6388800ba969db2045c9b28d103786012312e1085061008ad0b409da061210120212a76000394842901090906c1484210001004b4429c00866d44040d4429283462891406540165041a2858436328b2481c8c02d09970cccc808d1c86c1411100a182e193248c9c201a09840884072d3021023346ec9a41054006d53086420830c92440e21460688080d0a392201244d01936d0209710c29681b044d13b4241b2609db240513c52102124194166592c28d0b3188c1a251da168d8b388820c36840002503123212114a99962801372c8a182e63844818159041b491a3a64014214112b041d2188000970451366510b731dcb82110c20c613828d32448013389c4a02c6134000a454c50020d14270263342e42987123194811050d5a088a243471d1c62cc8a410d09641a23421093984130304604686132024141408d3160818008cd2a66019278a0a24698b362991b049219551e1a24c13c6855a18908bb851988210c4222ddc96450947210239928b9468131051d42601599841e2b85161988023976560480e183266a092494cc005d4246a08a424d438641b092584a65059c4859ab2208434221a882860146c9c366084167160128120c3315b08221a018812a2259c842d5bb06158c225d9a650ca082184a24008a58452420c99964cc924844a002d03096aa4046c908650943682d818659b140a0b208aa088000bc565122986d8944dd98844138681c418641934248192814c461014354d8c9830c4020e08425123090898a26ca328320936060b437140408c53448ee4486a08050824212ec140018b848c83a444e4805090308ee4b64921b3059444841a39725c323020b70dd916925b08044a8650020790244301e332851383880a3090d1120918021053946d122360cbb4010897250b1108938445d336722144089318228b0428d210280aa24018952193a82d03c08960c420d1248011a35101132ecb8009848405c886088308400c396421072d1a3765e4a6050c4705e2b82c50c861908025db206da2083203862d94142e00274c9420299804888b82508c24881044806120228216041a26619b223160288ad18430e126912043800095511a92889bc001a2120209a3052382099414719aa82021276c00020603116d9a346c5932866304091c398ddc2250c29044a39004e4c681cba681d4a04d522212cb0440a128091c318818379009b3505a127001964c434231e0044508168a00c84c18252524488e84c22d0a166e64b880a3c64953408e19810d43826d18458e229389ca204819454ad4824188b06d4c16640904625b809119976504b820d3a408224505e4c22d830420ccc88900388aca9871d9380a009448c9142ae184802003680238691bb80804438813c58009256d09c88dc2922c0ac245194101241730dc148c034371cb2408048891012244c9143280162601c431a044485a100441266523256142082e630249531232db146e93246e401630808621d338924c308921b5016382412283500ab88ce3040c371b6f07e5b71ef8356cbef9286af47fe05e852c2377c949b79484f87aa26a3c319d63c596a97e24c97663de07a819eb3dc79db5bffd2cf90da2d0351440bb2a0f1ffd23ac76e278b3c06b12901ef68b02d7f97b4f1604ea6269cdb934407b90f0f5d6358f05b37ab23696670b11464280509ddb4df4bc1fd5bccf536fe1ab908f39757994937d6aeb5596c45498813de84ef97f548c1dcf7144d0cdc455b08bf0d89ce32a59b12e81adc07e3cc22416778898f505b4bd0131d2ac7e01fe4e17eabfd4fe60523417db7c9207dcd43910a61e867133961484043bbd8374de9b07f38d5663a73e5e8d24ca1ae75dc304f7dc9629c1f7dbc819ce0cd93e3dc4ca62094dd51361e88533821bfaa77cd1b55049b373402918e4546b30918e150673ff833c64cfef2c58ebeeadc7d70d1524cbe71e02637562da42471551a22f7974b33a2ce80c385c3cc2ecae057a27f08e1c0298f8075ec1307bdb837d070dfea579a36022b3a809767271f2a2ee193989bc7e4375de3258c8a2ac84cc23b976751b3bc8b0e021b44df73d1908724726f215e021f8702fa60a204c636b4615fca78dfdba2767b52e30a8f5a6601fc5c4eae7ea2d1ceb705a2ce4a8f73d547abc2cd40fde20c152ae7c91c1230dd4057de4ed445c7f92af2153e1f78339f7e8b4ec953c45b164f26b46cac4d8bac7a1209785512dd9d3a4b4ab3d0b209eda7a081b6de0e43744244d10a1581245a0434c445bf111df890b2e3b4dfd4b313c46185c2c05b4ce7807471e9e33456abe7714f91cb9f7d73b13e4102bc97c62626d70fc53e2d9d9a8d1acf78549b18f65741fa132558ffb94405bc8934fb7b4c42193df9571d7f8f97fa86ee63cb515d45309cfbca0f7305be5f2a32aab74e7a7764a8bccd793bdd703cf5391d795d58d854c5bd5a6a4b3d164b2825d9c714c4c89552c5cf099d3d408bcfe74b74510f3a37029dbf87d863e4bdb8a782855d511526daf65d8c1a98c47fd99aea610e2fae1004c8c3263ceb29a968b9a5b246b94d6773415f96435859c7801df06b24a918d1e044ddfe38870c25bc7f4d4a61a5209ec2948ceb8f605fdb757d56d67441bcbab9b8091479a292925900cf8ec9f970987fecc7252ff889d04275eef9f54b4e3b72351bc7d1a41389db726c7235b955519930c22a5c99aab5431ef5f4f01f3b7c3cc78428610a26a1baed1123637a758c23d1f6654d87a655040577a16867c30a7da05b662c173b7f3c994a186f943446d89195a6797aaeb180abd5221c55533856bc92b9a6d1f382c4fbdf77b36d4bc26cd2106b6ce33451ddb7def9c235bae0b6bde13af9b05015cb9e6b605eda19a384222600ca8313c548c1383c7e30a607ca71e2870f4486f1d1987af632835edb055c59b84bd0848f2d6c97edacf4cd0eef76b89a48799354b8c15da8997b96a0b2d0f72a179f35d3560b012187da149e2110aeb548ee598fa5144c10683695e6d3564aad46ed88de5e9441cd15ca692d83ecc14242730b14bd2ca3a31cfbca6fb184ae165a48f527e2d4858c49ae0c018724ca49b52671ded5c93b2832d90351facf33a2f5c99325a62e4fd6c97321b07ccd3a6bb140c3e264b94292910b681c058836c7e8e60c570afd3c84eea41c6c76d3ea221c3ea5ed750ce70785f8795273edf7781f6f0ad450f170fea3c6cc2d0e7b4a5e825a4f4fb94dca3eb4b5392c870ac5aaf569391e654a52d2780b37668acf3870aa9040cdd57baee14df930b942643b0ec362875c5162466e3151d6e7eae759c4773be212f81579c463b51409df2b6790f4167ae46d82978982b52de58643f500a4567b6c8c82a50df8968809b2e3e3ad89233141ae014fa31597b6c54b0d9a82b2319108440c0bd9b85268fe14db4a00db15c8e24a3087863eaac98fb1c949ef4839d515ed4dd418ec4087e82a1b8bdbbdb286d09e16d04d12fdcc2cca4a40e9e320e729f29d4099ed88bbaeb80b5fbd141df323735e281ac104be541e21ea04d4efbaa24578bdc7f7dc9e7210df14efc79925ca631ec19c46a6c93bb6e1c565304a4ac055cfd8cd1fcdc65295a59f339510337b49619a54a0b3c7b4de7c6259891e3c444e3aa78aef5aa97392429827ccf40a2f34011e54faae2a066c5e4449948fe85b60e0bb2eb19b1b78a4d9f65d1d54f53b99a82eaaef195563bf7e05b38db7d41b38279cc525836b84214b91ebc3e17653e99b4470686f8bdfec12044ef3158c02fbf01aad0968ca715164646d1f25bbb00f57696d973cfd7185c0d5ea453ef88ef6eca42e7b7baa04515f2faebf488b5977ced64a9055bf2adfe46fff37ea7b376649d2f350dd75b619f36a12ccdb4b60036c0ef53f5e9e0616ff348c35a0e2170115cc8b71e94d31aa006cb9d1ed81edd118a0d813bb76df36398a0d112685780dd6c7a5325d6d4c67cb7114d372470b9c810a482bdc55cb275a45a514afa0c4a41d6b6393d61bf28ea66cb17c8fb4de5b7cfb56b1c8aad72654d519c8aaea675efe7c8d98e84e3ac936297f6c747182427876feb71865611b73459e0bc2ca1436980338cdc491756b5810ec194880f0ec9145cdd38310bad545d3998d84cd9523bc913ad410f369af1eb5b92824503ed6249048be6deaa7d5f44ea63df15290c322c4e86e76538e6cf965a25d2630283278329f673c8ebb19f62b97ab40fec7cee564b650e0cc255d528eea53706a245c98c530dced4493e4eeb4684b0881cbeff26b337c81e443be179085613b27e4d73456b61d899c60c848d9b5fb1a2f1a7125e7d3ef11f28e8aa8ec4673543af2bf4963507c5ec23627c2560e6a1be3668273cdc78bdfd7d4fb11adce565719d097ab9fc2d921e94673ce2f5859ec877e33b8535487218c63ade379e5c9ba1be1b92e4158ed7ec82b74c60a2cc7c4cb2c77eddfe8d2ccde1789b71723ab869d901c7c258fd5ff3a0cddb8378d85281f6fabc0cf1ddc8ed6c14622cafd9833d7a08659697796725f4aa27a7b767c0aa4172107311fdc9b9d57889a55e537f2e2a4af3240a3456a24475926aa305cb97ce73c0ac7a022c1dd21f0d9e32d0c38a2a5d74e2cee92aa2d290f81e5307916d6caf40c5620ea3670c2e2146ca3b2dd09a13a440a6d7febf5a89b018170e4bf2e8b36181cb67857b1013dc4eb59d757cc035a1f55703f28c3ecd9fe8c4ca4dbcbaee17462f6b940a0544fbb43e6fb1b19620d5e0aa84009cbe378e2eb8a4f7efbb55e1d47bddd11e49812b6574eb42e9d60fd2a97e7f7d77468ce8a923672b219b0546214cbe3eb5346f62526797b4d6bfb8878b53b18126a5fdc77ce07c8c529c83321d71f6dde838d93dc10d75a0210789166213191faf1897b80a5adb8cef9a68be4000029ca20ed5465d6193f7ccaebe7fa0777f4e1ae75f2662bbea0744b09d7e9d96544cde184caa940d303d1b4f50f10741b890983629c5498027c6858d2df3e936a9a6bed6e40237ce2bf90fe4d2de36639cbf0956c0de7748838b4f2f5fd99a47b4f2ee7570325bca5f58e8c2771d3cb6955d4610f605a339ba7f1e552f2403b07e4a9e91625b2403edf9416e5e0f03f419a63256d8f7b1d34f5e049bccecf2684a51f3600256f0e83dd7c94b3ff0a2c2df2744b6243b7162ba17ad3c5063a32060978e7fe1cc1ee704aa0cc78117e3741048969641c3c6c416bebc0d2e71f887a61586b3a875a84f416690ed0940f7f0f325d7686c357e1c3bf09f2e95ae89f1e588415eab9d9cfa9eec89e338f9672fe565fe89ac8f782a5635030ba27d69b1f4858592f29d798c19a62502d2bb1e4ff916a4cbd1f3896cbd4202acc219de5eaa861518376c26f4792ea841c2262be8923f79e6314b50e0b949ffc8a5a627a953036ea679a0f50ee1715e60c00090a4947674e27efd421d644472f99845d62448c093baeda9a246ab13f0083da70f7118c6951706bac8e83f47bcbd13913423c2b56b2428c4efa999952568f92a4ad3f36a79d3f6e2478e3906c9be90e0170d58065fe8c3ebe2f12bb91f9b6ffe75d26cc97a467a1d50490a5bf72ad2b4a3c74354ef304ca7747917f585ddaf2716ce620af5bdfa5062ecd94cc6ab1ba1923a21819e13fe4b5563e561e16a946ab068c445adc57d75fee0406d13bb2b28d466fe9c46c367c25496ae943800f7577196a3accad10c6ac9547bf7775ce388e9806e15c534431a75d3d8b301b553ffe94feac0a26da61763bbce647fc00b439e25c5966640f3426a9eeff1128f4619f48ae7ccf0a74259efb01d294e5aacae3cc426ced3f7bf743c592cb25879ebc35d470383c13395d43076a48399c5a2c9c2f1086367c14f33bafe68cd007b3d65efde8698d3b18db270dd3321774052799754e861e9b82ed20f31dd9ac0086cb50466e6a1ccc87d9c02570ed5b1aed39fbaf1190c1696118d0ed21e15bd1aca546636985d3b65a6a1ea689bd20365ee6fda52f8bad51fd3344defd450296bfc09844bedd57399b7e6dc956fad723d4a187b09e2b3f215d50a11c7f93c594b8bff1ba94bfbea7e05c35eb9e32ab6cbee4f90299bbb88b019b5819ac0189b7e09b9365cce29b9825b42bb1a6cde57e97f921300e29b7bf7f6c0b915e91751ed1b3d5e4cb0f58b730b22e197a0853c3a82bba396b86e6ac9c0981cf3a3b4e2389310627aeef1d289cefb28bcbe0d6adb4390e288ed76b10ae7a20b",
-    derivedSig: "3be701f504264b93ea229ec1179a6e82b15257f12ddfa312b431f0612bb8798d0aff7ac34cccf70084f03e18f7774f66d501312e72f1d82f936bbfcff8ccbd394b0ec97c339535d3cc02d2d0883a50014a48bc6987bcac563178fd7f56478f72363926b4ae2ff03e1e1c73797b4e0276ace553d39fc6658e395e0f6526b91e0453d4705c18305b264568f10b6b3378343059232c393865f4f6d168eb767419eb91da75717ef7e11a52c61d7387ef12081b2c50f491e25576fa19af8fd23ed1327a74c0b942fa8015dedfeef072006da33576a28ac40692fdb668a33f070e0b3d516b94799f0a0d7d1d27be1ba5c1b5eb394c323d2e2976d6866ca85240410a4cbe28062dfc11f3d22640870202f76fbba145d98cb3bbc356e414f00d3b965cda31c4b3aba939c2eb629ca6dfb3e9c429591feec2193ebc94e0447bb3df2499789bb0231cbf5729ba33ab58a2726f8b51aa5e35a88c683911ba6a3df8019370ee03c09d1fc5bb74d2a64e5272bb64ba26cac9c898b8a381f1acf781734511f755466aefa8e4db05164af440ecc1f8aa340e6acdcd894c94cc2cc76a45fb16eba115d5dc0505bcebbab5578dcf220ff4afacdcb0999413240857201359feeafe8bbe2e20b14dcdecd551685ecf6a45f716de2667d2db2c2564704fcd5d514f22f7971e297948224139460f0a6dd1b25a1da9cd3cfd21d722ede6db96f88c66af837a9d0ece1f2b3f854a85b1141453c3609904b68bb9dd6190af5d1e21e3c6d315ee1185c886aa59f6328b27f4e19d74c95b59988889ffa186bf976f3f60c99466fb5cc8231399b94fc3970da9b2be81af34c046b0d562276138ee3c7ddc0ef07567259c47c89f80e1f29f9679a5aff721b79c665d74395959ee37072bd31d75c482f62b0e1313f9121555e3807cc69f7692acfc2958765dc618561733a709e3218c797f3242c1ceddf9cd394ef80c4e4cb8765db6352d2d8fa62af7043b6f1995bf72887b793f25adf0674a8a95a285ca87e5cbfc2e9834513d9f48822b7e7c4d70ca878d267851be95f66bb9ad19a02a8061bf300cbf9b9bc9a4745d3f212ed0486028f2990008bad005155af6589bd0dabe9e1ae8cdf65e9d6fc8e0032739b444fa7ac4aa063d5be36b115a9c262a48cd17379131944b0e3254572cd3350d915004aafa8f46626a037d4222995a4c2ca2dd0e566176f80d72f2000c8fb4b34d2585433f04f9e51b1d62f3c5438c18793da0521386b78282b123bbe7852034cc168266f7efcc222c7c6caf0d43618be06465948eb3f7c4068e92c7583850303bdf439b7155a60ca6d4b4892655d8b6bcefbfbcec07ab1d69055613a2a0ac7f5867b1083b56a0338efa128f2d404b531e86c03a54e81f3ee89aaa519fb3812bef398b24c8e98c4014e70a233fc88a33a2341b97e656c755e9436dcabe3691463b5fd2724b4c8d7a18a6938c74570f7b3091511f0271c03d35975612f910c71be84aa4718a5d2bfefad3d95d064701369f103e0b91e6174a83f78d7a2ef7d32f5cef60d2f2a7dbbc0eb1e72566c9cd68b12d86a00f0fa4362ecd753ed2a55f84b31aaa0b716ad8d2985a424a80cfde5f1d8790db86ad24dfb982c730b38b6ded3d34c8c55aaa787d6283beae7e93d7184a7f606d810745c3c7fd75079f78dc464a6edfce293608ccdfde1da0213339466f37f5051ba1417f55e1881d3bd622afdc2b2d3d27e3bbcaf5c2752b24f6aa271b2c2c3a19ae8a7aa7e2b8a3ac40c29931949487dd3917fbe743e08d37a375b7c0a2cf902ccee607092ee67fb2e70de9f38f49893df2d565884e02c0e59e720c1d481a178ed990f93556f6e6c6ce7ccac2c0d34a2cf295e200fe420bdb40f999c799e3beae387d7c6586a909324ba5b69d52a3d47430548e7318e013d68fe08194af1cbfcb61fabd30041c729585fd0fe995a38cac5a3142196f1212b9c4035f289d60632d43645fdfb2bafcf273bcab9bc9a6fd8c694feba54b8b25260b16dc0e1bc6aae72bf3604ccd526bfd5f68b6c26f74c6207232e6f73308b89fd431d5a5b40c396b0271592acee81eac728707f29a3424ba6ca394594268ddf634cfabed5d0c2cc46f15703c2826b4a59c0c23e63b9150903f00ec53d853d7f11faace87e6bf5f1c329889ab548668f964c20663f7e46fd96c2e6d0b52d2a85a1c193b21fedab5aa24887385745f6f9fa5d886d29701d6f5f81ea3998334440e9dc86b6b680513d562a0146e232cdb306840229ee029086b325aba81ee8c1f4c1dc09d35a853805a434295b71e5ec220f2d7255ea16da789b0573362c3dd12315e0ecf84a5e5f7f73a9d8ce47b3c510c52f7a35bb5135598c15f23da20b234a648f73891ebab3f1e5faf177470a80d58ab5c546350ee6d97cac210a98a286be9bf028c085023729da8b91c23a91408cda7c003e0b11e97a26b31503ff03b59577b2082cffd98c8c6b20870ff7c76014a29a8d0b991448570c6d29958e08d93ff636522018da5f709e8de0bdd7aab278e3f795c8a115c92befeabb1681e7f0f5593a2f88fbaba49667b3acff8772f08c0cecce71f348c7fbc9c3ddc5b408e1e9f79549406c13c4fabd8b645953b393e8f4b8370598f3b272a19f37da4442e83b39c88fdc6bb8c9df0c0eb47fef920830f6659d7e603f62391606076182fc0271f98ec29c178dda610dc92fe35e2510b7eebd0933eef991d902a72e6ce88e10efb1ac575852869b204b9773f6d1eaa492cefaf910531f5970ad0db69d8c064ebde401b8477d6081b9edb914709573984393da32d40b159655e07c3333f108996ba032c4ec8661ecd53d0a1b6d125308c1ede849b810e94fae380c50ee26ccb5f88939529008f3253e7ebd984f99aba2f2a749e617cf64eea03548291701e45ce7357fe8dbf91154dc2c4f8a17759b2c05c87e16d4ee4ec56a9d116e95b84ec3a01dab692fc52b1c56658ecb153008268df898fbbeb98bcba5e91380f078a2171c9ea86e5cca79b996d563acdd4eac81f10c196a4deaa499ed2b8205713a9522381c1f751a579dc1ec7a458c3756a67c719253a83c21753361b9408a3d8bd58b96d25d78b99350537a43804f81bd46f3da8ee7479f87ee0d611e3dc8aa2b651edccd83eff9a5e74d4413e8f274dca5388e9ff5a6d6aabeaf1901aa14f69d96c1672f61d313f59020908e8863ef23b41cacc444a7595adea00975e3e3f7fc9219093ec3acf88efc59fcde079fdcc3345819594ff93098a53c6e25b72cc78f0faa1f28b04cdd0103e1aa80a03726f6b4c0a5755dfdf4040b4c8e952350db1de23b6ade29a8c2e58b7e10c82a7875afce516989b458a43dc888183264031a9a86d29b015ed39b1ca2f0a705b877619b787bdff85422e587834b160a7e35146f7baa3fd0723afaa179d79a008c6afbe445255c3b1f729808adbb8e69f06691262945fba7ec026afcbdc12e4fb4388f88804ab1aa9745c97cc886ba71ef2a85f6d62d5c50d80dd89627c1a46fce42455b5bfa5c439abb7c97e5826954245e9d9390913d183365df3019affc5a9164cd4058ea171c8c86913534fff330c2941f7bc0718884a84c55a3a726d9334ca5514b02c597fe2061426280a08a2ae22ad693cee708707d23aebf3b9927f87f76c0f6aa4c5c10b6bd1b738825904d86bbe95a894d0af5fa963abf9acd3dc28c09a1ffec0eea1cec3aa0bfdd4e5cbd0ea4658f10b4f45e12c02f836c3f6ddacbd96d5713bd226dce3b84925d60b4d6cbf3cedbbbe61b8250749c2888ab75624b0d66fbb4c5d4b674ea75290630a06b404f1a3f9d556ea3581ee954b0720d1a2720c955eeedceaff5e595c3e0d9ad157eab1a591f01d705816fa8d0ab08f2e739b66a9c4f35988808a09de4f7ce3ae757e83126cca2e5d73dc0249d1073e53fce24833743e722f55275f5e796f48605b7724839249ca41fe35d3da90f570a0f957da4ed9dd25b697c313843cea1792c3fc8f96bb24cde438e8ff15580771ba7de7557f7946f776ad042e6c4dec9fc93a99078726ae782d8a1c85140ae7ceec46551e4a408f6cfa1ea09c0807eff0d192272faf91d8b47436d9091755140da09fccd33218a5e3ce16818f2a36b6b4f563112d3edec6cadcae7f88fbdd1c260122e93b2b9da595d1bc7b6df954ab99eb44de14d91333590fde756fa255581db1105720691fecad50a7a438b7d25cad5cabdfcee77f816de76bfbf54c186924020260341a2e43f7cf0a14f06b7a08c92f5753e8239fd505d59033f648aee8a5a94fd9777e5be9ce77a66417c79cae27919ea205005589ab37170df4e37079eea187e9a5fcdcd44d5f72f9fe4498906ef04d85cd6f8d0c36cb65d30afb6fb495ef7559c90135cd9304f49362e6b0d8678b5a7eb72944a65fc28c844814a742707dc729361227aead38184b425c26c9f8c9a7b1d12ad00d231302f3d6fccf337dc31d2f7ff4a698fef4b55bbe1f59230dc6902869d6268164ed12faf3795a52e5e22f55e16b290e8731546b2d686cebfa32d9e4494b02e2fa437cf02bd613fc4cd7c1601799c7a95d41b4880ebf3ce001f4492eff12cda040e53ef4e95b1222d289a2c60ec90258cafb7f401f32de4390b46282f9bba35e51eb21fbad44ba8f807fcccfced9546327912dc52313227608bbc1a948b4188c501e05d12eb1cb9628ce9199c4009a9314ef777157da03bd902d88cc3c7f9442bc0712df5c9918601c5007e4dbb45ef2751d30b96795bc8c64c04dafe08559e96199906a55e96f04d844cf4769ee551802e27bb1d9bba82f99fef6eb9769bea0c641dd961743b729150915b54296777cd46ef6a50b8917e2655a19175cf74df28df6e3159c0a7dc2b709b4d3e53b0ebb30816172a10aa0d04089ac639401e89c58ca399b1081ebc1dbde0f69a44724374b0a34592123b5329b520856ef420dc4ad1d158bab469825ce5ec903cb4c899ad5f7714d0d030a261588ac9e2e30ba79d9e5d62b6e07990b753783de4693eb55357882538c7ccb5695098dc65ef55f4fbeead5d8108a62d0aa49c51a538641e2eba8c8e97f0000d526bc52a77f6f22d86ac9eca2ee97ded2d6a06e5f240474131afd674b9052ea95751278bfc4e6d020ed490c1ec9a587a8ebd31dd19d36612fba50d5d1d6e35f1ed2dbb45069dedcedadca12f917642658e6f69ae9a6de1686b80094e724629b93f9feb6cbd65c1099eb63582b80d5a6d9d26466dee5c09aea183098b5011401af63a46fce991f72e469e3e5798c99d18c6ce1211d2cc53807669815494e57f81a21250cf0676ca9d63b15d317f22e487abe4918e7302a8eee876cdf8411618545bf2d4851b394663c418b6a05658a469f89dc8268cbc9696bc6ed92ba01ccff30f9752792f41f323676442ab00d2bdc1b968aa67ba3e9d8ad1745844bedd737cf42f428e896d26e235d09736548b135b3f0d4811db639108c06c5127be8ab9b2ddd4e882a1ee240be08aff1fe293b010afeb02e58c8fffc97c9a06473257eb460aec5ffe95820004c25da81cf2a304e498eb0d58c60b5ea257ef51d2b06e6d8302202bc826e7ee5cf6a9428435a8fe2ee84301297b41ceaff5cc629c16bf8d63dd9a64d054d4a36cc592c93e62a4d384e0cd3b5167dff6b468ae0039e353858c20fe5f0bb64c7447871a36b7e8c50ffc2c04cf6496a7e4f0e138716538f23f77433f54944fda8bff6fbddee9391fc0a19276cf474afe66cc933084ed99ab5d0888e215f3558bececdd70be37cdcf2d3a61cf26521ff301e5044a2bcbd8853be260bf4881bd6c3a67d4f1532fc9bc1b35f4fe7e5d3d0f80675b8e370a18f4b9e656eb7ff288645c1a18b31f3cac9fa344556cff78ac615a4bb6944b7974c415ae796d180fa479c45b017973a5cf8741f3a65472993db60fa30c517395dae296762416173f9e62de543488eb891560be885e9b6125e21630a4b955b4192e04d120a21b77008316c5c58c7a411a73fd87d5101800b8b10ce50f67df849f1339ffd9503ac2b5da40d63287320c36296fc6d4a626ccf803b6b356faf73a3ad54a4d82f20a4fa0c5c6302723e7f5e330b9bb218eb80e10fe75c6569820fcf8b781a19cb4646f66f4feb11db11cf5d3b114b55d9093a8504dc8cde4bcf259cefe2b5d6d5f91ae5a7fec173541d751492ad7d11797986c87dcdb1b74b205777d861eff072670a68c6856bb6edaacb9503fa97471c54a7128a826ba82f9ec7df0ae4ddb88adf88559395deea942a233b848dd56bd6e9d34f246012a9515632bd63282aa2772f4881781bc353a59e06aa5accad4616878a7bba8014755386521e7eec2d467aaa21b40a0304343b8d90d12d39d478a2eba5e24df7c11f055bf46a182c576b034c7924f069f1a0723785d37b1fe4d41e209bd9b3ca255fe8a72eafc450511f457b85b7dbf8db011329406db9dee5f8fb1f88919bb0526db6c6cfd5daf8fb5e758492a60883c5d80016243244a6daf000000000000000000000000000000000000000000000000000000708121720252931",
-  }
-]
-
-describe('cryptoSignKeypair', () => {
-  for (let i = 0; i < TEST_VECTORS.length; i++) {
-    it(`should generate the expected pk and sk ${i}`, () => {
-      const vector = TEST_VECTORS[i]; 
-      const pk = new Uint8Array(CryptoPublicKeyBytes);
-      const sk = new Uint8Array(CryptoSecretKeyBytes);
-
-      cryptoSignKeypair(Buffer.from(vector.hashedSeed, 'hex'), pk, sk);
-      expect(Buffer.from(pk, 'binary').toString('hex')).to.equal(vector.derivedPK);
-      expect(Buffer.from(sk, 'binary').toString('hex')).to.equal(vector.derivedSK);
-    });
-  }
-});
-
-describe('cryptoSignKeypair with null seed generates pk/sk', () => {
-  it('should generate the expected pk and sk', () => {
-    const pk = new Uint8Array(CryptoPublicKeyBytes);
-    const sk = new Uint8Array(CryptoSecretKeyBytes);
-    cryptoSignKeypair(null, pk, sk);
-    expect(pk.length).to.equal(CryptoPublicKeyBytes);
-    expect(sk.length).to.equal(CryptoSecretKeyBytes);
-  });
-});
-
-describe('cryptoSign', () => {
-  for (let i = 0; i < TEST_VECTORS.length; i++) {
-    it(`should return the expected signature message ${i}`, () => {
-      const vector = TEST_VECTORS[i]; 
-      const msg = Buffer.from(vector.msg, 'hex');
-      const sk = Buffer.from(vector.derivedSK, 'hex');
-      const ctx = Buffer.from(vector.ctx, 'hex');
-
-      const sigMessage = cryptoSign(msg, sk, false, ctx);
-      expect(Buffer.from(sigMessage, 'binary').toString('hex')).to.equal(vector.derivedSig + vector.msg);
-    });
-  }
-});
-
-describe('cryptoSignOpen', () => {
-  for (let i = 0; i < TEST_VECTORS.length; i++) {
-    it(`should return the expected message ${i}`, () => {
-      const vector = TEST_VECTORS[i]; 
-      const sigMessage = Buffer.from(vector.derivedSig + vector.msg, 'hex');
-      const pk = Buffer.from(vector.derivedPK, 'hex');
-
-      const openedMessage = cryptoSignOpen(sigMessage, pk);
-      expect(Buffer.from(openedMessage, 'binary').toString('hex')).to.equal(vector.msg);
-    });
-  }
-});
-
-describe('cryptoSignVerify', () => {
-  for (let i = 0; i < TEST_VECTORS.length; i++) {
-    it(`should return true ${i}`, () => {
-      const vector = TEST_VECTORS[i]; 
-      const sig = Buffer.from(vector.derivedSig, 'hex');
-      const msg = Buffer.from(vector.msg, 'hex');
-      const pk = Buffer.from(vector.derivedPK, 'hex');
-
-      expect(cryptoSignVerify(sig, msg, pk)).to.equal(true);
-    });
-  }
-});
-
-describe('cryptoSignSignature', () => {
-  it('should throw on invalid sk length', () => {
-    const v0 = TEST_VECTORS[0]; 
-    const sk = Buffer.from('00', 'hex');
-    const msg = Buffer.from(v0.msg, 'hex');
-    const sig = Buffer.alloc(CryptoBytes);
-
-    expect(() => {
-      cryptoSignSignature(sig, msg, sk, false);
-    }).to.throw('invalid sk length');
-  });
-});
-
-describe('signingErrors', () => {
-  const hashedseed = TEST_VECTORS[0].hashedSeed;
-  it('should throw an error if the pk or sk is null', () => {
-    const pkBad = null;
-    const skBad = null;
-    const pk = new Uint8Array(CryptoPublicKeyBytes);
-    const sk = new Uint8Array(CryptoSecretKeyBytes);
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pkBad, sk);
-    }).to.throw('pk/sk cannot be null');
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pk, skBad);
-    }).to.throw('pk/sk cannot be null');
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pkBad, skBad);
-    }).to.throw('pk/sk cannot be null');
-  });
-  it('should throw an error if the pk or sk had an invalid length', () => {
-    const pkBad = new Uint8Array(12);
-    const skBad = new Uint8Array(12);
-    const pk = new Uint8Array(CryptoPublicKeyBytes);
-    const sk = new Uint8Array(CryptoSecretKeyBytes);
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pkBad, sk);
-    }).to.throw('invalid pk length 12');
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pk, skBad);
-    }).to.throw('invalid sk length 12');
-    expect(() => {
-      cryptoSignKeypair(Buffer.from(hashedseed, 'hex'), pkBad, skBad);
-    }).to.throw('invalid pk length 12');
-  });
-});
